Database seeded

+++ db/migrate/20210423225341_create_bulk_discounts.rb
@@ -2,7 +2,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
       t.references :merchant, foreign_key: true
-      t.float :percentage_discount
+      t.integer :percentage_discount
       t.integer :quantity_threshold
 
       t.timestamps
